requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
boardgamegeek2==1.0.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Deprecated==1.2.10
discord==1.0.1
discord.py==1.3.4
idna==2.8
multidict==4.5.2
pycparser==2.19
PyGithub==1.51
PyJWT==1.7.1
PyNaCl==1.2.1
requests==2.21.0
requests-cache==0.4.13
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
websockets==8.1
wrapt==1.12.1
yarl==1.3.0
